# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "post": {
        "description": "Crea un usuario genérico en el sistema.\n\n**Nota:** Este endpoint **NO** crea automáticamente el perfil de `Doctor` ni `Patient`. Si deseas registrar un Doctor o un Paciente completo, usa los endpoints específicos:\n- `POST /user/doctor`\n- `POST /user/patient`\n\n**Roles permitidos**: ADMIN.\n    ",
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "firstName": "Juan",
                    "lastName": "Pérez",
                    "rut": "12345678-9",
                    "email": "juan@climas.com",
                    "phone": "+56912345678",
                    "roles": [
                      "ADMIN"
                    ],
                    "isActive": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o el RUT/Email ya existe en la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "El RUT ya existe",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo usuario (ADMIN/General)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "firstName": "Juan",
                      "lastName": "Pérez",
                      "rut": "12345678-9",
                      "email": "juan@climas.com",
                      "phone": "+56912345678",
                      "roles": [
                        "ADMIN"
                      ],
                      "isActive": true
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los usuarios",
        "tags": [
          "Users"
        ]
      }
    },
    "/user/doctor": {
      "post": {
        "description": "Crea de forma atómica un **Usuario** y su perfil de **Doctor** asociado.\n\n### Detalles importantes:\n- Debes enviar el `areaId` real de un área existente.\n- Retorna la información completa, incluyendo el `id` del registro `Doctor` (este es el que se usa luego para crear `Consultations`).\n\n**Roles permitidos**: ADMIN.",
        "operationId": "UserController_createDoctor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario y perfil de Doctor creados exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "user-uuid",
                    "firstName": "María",
                    "lastName": "González",
                    "rut": "98765432-1",
                    "email": "maria@climas.com",
                    "phone": "+56987654321",
                    "roles": [
                      "DOCTOR"
                    ],
                    "isActive": true,
                    "doctor": {
                      "id": "doc-uuid",
                      "medicalLicense": "LIC-001",
                      "area": {
                        "id": "area-uuid",
                        "name": "Cardiología"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o el RUT/Email ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "El RUT ya existe",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno. Posiblemente el `areaId` no exista en la tabla Area."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo doctor",
        "tags": [
          "Users"
        ]
      }
    },
    "/user/patient": {
      "post": {
        "description": "Crea de forma atómica un **Usuario** y su perfil de **Patient** asociado.\n\n### Retorno:\nSe devolverá el `id` principal del Usuario, pero también el sub-objeto `patient`. El `id` dentro de `patient` es el que debes usar para registrar **Consultas Médicas**.\n\n**Roles permitidos**: ADMIN.",
        "operationId": "UserController_createPatient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario y perfil de Paciente creados exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "user-uuid",
                    "firstName": "Carlos",
                    "lastName": "López",
                    "rut": "11223344-5",
                    "email": "carlos@climas.com",
                    "phone": "+56911223344",
                    "roles": [
                      "PATIENT"
                    ],
                    "isActive": true,
                    "patient": {
                      "id": "pat-uuid",
                      "dob": "1990-05-15T00:00:00.000Z",
                      "address": "Av. Principal 123",
                      "bloodType": "O+"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o RUT/Email ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "El RUT ya existe",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo paciente",
        "tags": [
          "Users"
        ]
      }
    },
    "/user/doctors": {
      "get": {
        "description": "Retorna la lista de todos los usuarios que tienen el rol de **DOCTOR**.\n\nÚtil para poblar selects en el frontend cuando se va a agendar una consulta. \n**Recuerda**: Usa el `doctor.id` devuelto en esta lista para las consultas, no el `id` del nivel superior (que es el del User).",
        "operationId": "UserController_getDoctors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de doctores obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "user-uuid",
                      "firstName": "María",
                      "lastName": "González",
                      "rut": "98765432-1",
                      "email": "maria@climas.com",
                      "phone": "+56987654321",
                      "roles": [
                        "DOCTOR"
                      ],
                      "isActive": true,
                      "doctor": {
                        "id": "doc-uuid",
                        "medicalLicense": "LIC-001",
                        "area": {
                          "id": "area-uuid",
                          "name": "Cardiología"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener todos los doctores",
        "tags": [
          "Users"
        ]
      }
    },
    "/user/patients": {
      "get": {
        "description": "Retorna la lista de todos los usuarios con rol de **PATIENT**.\n\nÚtil para agendar consultas o buscar pacientes.\n**Recuerda**: Extrae el `patient.id` del objeto devuelto para vincular consultas, no el `id` superior.",
        "operationId": "UserController_getPatients",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de pacientes obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "user-uuid",
                      "firstName": "Carlos",
                      "lastName": "López",
                      "rut": "11223344-5",
                      "email": "carlos@climas.com",
                      "phone": "+56911223344",
                      "roles": [
                        "PATIENT"
                      ],
                      "isActive": true,
                      "patient": {
                        "id": "pat-uuid",
                        "dob": "1990-05-15T00:00:00.000Z",
                        "address": "Av. Principal 123",
                        "bloodType": "O+"
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los pacientes",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/user/{id}/disable": {
      "patch": {
        "operationId": "UserController_disableUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario desactivado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "firstName": "Juan",
                    "lastName": "Pérez",
                    "rut": "12345678-9",
                    "email": "juan@climas.com",
                    "phone": "+56912345678",
                    "roles": [
                      "ADMIN"
                    ],
                    "isActive": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Usuario no encontrado",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar un usuario",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "API CLIMAS",
    "description": "Documentación API CLIMAS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "rut": {
            "type": "string",
            "example": "12345678-9"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "johndoe@climas.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "MiSuperPassword123"
          },
          "phone": {
            "type": "string",
            "example": "+56912345678"
          },
          "roles": {
            "type": "string",
            "items": {
              "type": "string",
              "enum": [
                "PATIENT",
                "ADMIN",
                "DOCTOR",
                "WORKER"
              ]
            },
            "example": [
              "PATIENT"
            ],
            "enum": [
              "ADMIN",
              "DOCTOR",
              "WORKER",
              "PATIENT"
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "rut",
          "email",
          "password"
        ]
      },
...
    }
  }
}